// src/modules/com/cliente/dto/update-cliente.dto.ts
export class UpdateClienteDto {
  int_codstscli?: number;
  int_codtipcli?: number;
  vch_cgccpfcli?: string;
  vch_nomcli?: string;
  chr_tippescli?: string;
  vch_doccli?: string;
  vch_nomrdccli?: string;
  chr_cepcli?: string;
  vch_estcli?: string;
  vch_cidcli?: string;
  vch_baicli?: string;
  vch_endcli?: string;
  int_numendcli?: number;
  vch_cplendcli?: string;
  vch_telcli?: string;
  vch_faxcli?: string;
  vch_celcli?: string;
  dat_nsccli?: string;
  vch_crecli?: string;
  vch_nomrefcomcli_1?: string;
  vch_telrefcomcli_1?: string;
  vch_nomrefcomcli_2?: string;
  vch_telrefcomcli_2?: string;
  vch_nomrefcomcli_3?: string;
  vch_telrefcomcli_3?: string;
  dec_vlrrndcli?: number;
  dec_limcrdcli?: number;
  vch_obscli?: string;
  dat_datpfrvnccli?: string;
  chr_sexcli?: string;
  int_codreg?: number;
  int_codcid?: number;
  vch_endentcli?: string;
  vch_endcobcli?: string;
  chr_flgmtzfil?: string;
  bin_codclimtz?: number;
  vch_codmtccli?: string;
  vch_numbco1?: string;
  vch_numage1?: string;
  vch_numcntcrr1?: string;
  chr_digcntcrr1?: string;
  vch_dtlcntcrr1?: string;
  vch_numbco2?: string;
  vch_numage2?: string;
  vch_numcntcrr2?: string;
  chr_digcntcrr2?: string;
  vch_dtlcntcrr2?: string;
  vch_numbco3?: string;
  vch_numage3?: string;
  vch_numcntcrr3?: string;
  chr_digcntcrr3?: string;
  vch_dtlcntcrr3?: string;
  chr_flgterasncli?: string;
  dat_datterasncli?: string;
  vch_desitvcli?: string;
  int_codemp?: number;
  int_codmtvstscli?: number;
  int_numdiapodcli?: number;
  int_codusr?: number;
  int_codusrpcp?: number;
  vch_orgemsidtcli?: string;
  dat_datepdidtcli?: string;
  bin_codptucli?: number;
  vch_ceinsccli?: string;
  dat_datemsceinsccli?: string;
  chr_estceinsccli?: string;
  vch_livceinsccli?: string;
  vch_folceinsccli?: string;
  bin_carnaccli?: number;
  vch_nompaicli?: string;
  vch_nommaecli?: string;
  vch_insmnccli?: string;
  chr_tiprecimpcli?: string;
  int_codati?: number;
}